django-backend/requirements.txt
asgiref==3.7.2
Django==5.0.1
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==24.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
